FIX: Full project autopsy — 7 critical/high bugs fixed
Bugs fixed in this commit:

1. ShizukuManager.destroy() — Listener removal broken (CRITICAL)
   - Used empty lambdas {} which created NEW listener objects
   - Original listeners were NEVER removed, causing memory leaks
   - FIX: Store listener references as properties, remove exact same objects

2. gradle.properties — JVM memory too low + wrong Java path (HIGH)
   - Xmx900m caused OOM during assembleRelease with Compose + Rust
   - Java path pointed to java-21 but build.gradle.kts uses Java 17
   - FIX: Increased to Xmx2g, fixed path to java-17-openjdk-amd64
   - Also enabled parallel builds and nonTransitiveRClass

3. JarvisSensoryService — VOICE_RECOGNITION causes system beeps (HIGH)
   - Used AudioSource.VOICE_RECOGNITION which triggers beeps on Samsung/Xiaomi
   - FIX: Changed to VOICE_COMMUNICATION with MIC fallback (zero-beep)

4. GestureController.dispatchGesture() — Always returns true (CRITICAL)
   - Set result=true/false in async callback but always returned true
   - Callback result variable was never actually used
   - FIX: Return dispatchGesture() boolean result directly
   - Also fixed pinch Shizuku fallback: was two taps, now two swipes

5. RoutineEngine.evaluateTimeRange() — Overnight ranges broken (HIGH)
   - '22:00-06:00' would always return false (range 1320..360)
   - FIX: Detect wrap-around when startMinutes > endMinutes

6. RoutineEngine.getAllRoutines() — Gson NPE (HIGH)
   - gson.fromJson() can return null for JSON 'null'
   - FIX: Added null-safe operator with emptyList() fallback

7. HomeAssistantBridge — inputStream crash on error responses (HIGH)
   - Reading conn.inputStream on 4xx/5xx throws IOException
   - BufferedReader never closed (resource leak)
   - FIX: Use errorStream for non-200, add .use {} for auto-close

8. DeviceMonitor — Integer division truncation (MEDIUM)
   - usedPercent = (usedMB * 100) / totalMB truncates to 0 for low usage
   - FIX: Use (usedMB * 100.0) / totalMB for floating-point division

+++ app/src/main/java/com/jarvis/assistant/monitor/DeviceMonitor.kt
@@ -78,7 +78,7 @@ object DeviceMonitor {
             "totalMB" to totalMB,
             "availableMB" to availableMB,
             "usedMB" to usedMB,
-            "usedPercent" to ((usedMB * 100) / totalMB),
+            "usedPercent" to ((usedMB * 100.0) / totalMB).toInt(),
             "isLowMemory" to memInfo.lowMemory,
             "thresholdMB" to (memInfo.threshold / (1024 * 1024))
         )
